requirements.txt
flask==2.2.2
connexion==2.14.1
pyodbc==5.3.0
python-dotenv==1.2.1
werkzeug==2.3.8
flask-sqlalchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.26.1
sqlalchemy==1.4.46
requests==2.31.0